refactor: Move crud calls to own lib files for redis and postgres
